Clearer formatting, update links
It's now really easy to make it a Markdown file -- just rename it to README.md and all the headings will show up in bold :D 

Also, it's clearer even without Markdown, so you can keep it as README.txt.

Links used to be code.google.com but they redirect here. Who knows how long that will last, so might as well update the links.